+++ sctoolbox/plotting/highly_variable.py
@@ -1,14 +1,19 @@
 import matplotlib.pyplot as plt
 import sctoolbox.utils.decorator as deco
+import sctoolbox.utils as utils
 
 
 @deco.log_anndata
 def violin_HVF_distribution(adata):
     """
-    plot the distribution of the HVF
-    :param adata:
-    :return:
+    Plot the distribution of the HVF as violinplot.
+
+    Parameter
+    ---------
+    adata : AnnData
+        AnnData object containing columns ['highly_variable', 'n_cells_by_counts'] column.
     """
+    utils.check_columns(adata.var, ['highly_variable', 'n_cells_by_counts'])
     # get the number of cells per highly variable feature
     hvf_var = adata.var[adata.var['highly_variable']]  # 'highly_variable' is a boolean column
     n_cells = hvf_var['n_cells_by_counts']
@@ -25,10 +30,14 @@ def violin_HVF_distribution(adata):
 @deco.log_anndata
 def scatter_HVF_distribution(adata):
     """
-    plot the distribution of the HVF
-    :param adata:
-    :return:
+    Plot the distribution of the HVF as scatterplot.
+
+    Parameter
+    ---------
+    adata : AnnData
+        AnnData object containing columns ['variability_score', 'n_cells'] column.
     """
+    utils.check_columns(adata.var, ['variability_score', 'n_cells'])
     variabilities = adata.var[['variability_score', 'n_cells']]
     fig, ax = plt.subplots()
     ax.scatter(variabilities['n_cells'], variabilities['variability_score'])