+++ sctoolbox/utils/adata.py
@@ -1,5 +1,7 @@
 import numpy as np
 import scanpy as sc
+from collections.abc import Sequence  # check if object is iterable
+from collections import OrderedDict
 
 import sctoolbox.utils.decorator as deco
 from sctoolbox._settings import settings
@@ -150,3 +152,65 @@ def save_h5ad(adata, path):
     adata.write(filename=adata_output)
 
     logger.info(f"The adata object was saved to: {adata_output}")
+
+
+def add_uns_info(adata, key, value, how="overwrite"):
+    """
+    Add information to adata.uns['sctoolbox']. This is used for logging the parameters and options of different steps in the analysis.
+
+    Parameters
+    ----------
+    adata : anndata.AnnData
+        An AnnData object.
+    key : str or list
+        The key to add to adata.uns['sctoolbox']. If the key is a list, it represents a path within a nested dictionary.
+    value : any
+        The value to add to adata.uns['sctoolbox'].
+    how : str, default overwrite
+        When set to "overwrite" provided key will be overwriten. If "append" will add element to existing list or dict.
+    """
+
+    if "sctoolbox" not in adata.uns:
+        adata.uns["sctoolbox"] = {}
+
+    if isinstance(key, str):
+        key = [key]
+
+    d = adata.uns["sctoolbox"]
+    for k in key[:-1]:  # iterate over all keys except the last one
+        if k not in d:
+            d[k] = d.get(k, {})
+        d = d[k]
+
+    # Add value to last key
+    last_key = key[-1]
+    if how == "overwrite":
+        d[last_key] = value  # last key contains value
+
+    elif how == "append":
+        if key[-1] not in d:
+            d[last_key] = value  # initialize with a value if key does not exist
+
+        else:  # append to existing key
+
+            current_value = d[last_key]
+
+            if isinstance(value, dict) and not isinstance(current_value, dict):
+                nested = "adata.uns['sctoolbox'][" + "][".join(key) + "]"
+                raise ValueError(f"Cannot append {value} to {nested} because it is not a dict.")
+            elif type(current_value).__name__ == "ndarray":  # convert numpy array to list in order to use "append"/extend"
+                d[last_key] = list(current_value)
+            else:
+                d[last_key] = [current_value]
+
+            # Append/extend/update value
+            if isinstance(value, list):
+                d[last_key].extend(value)
+            elif isinstance(value, dict):
+                d[last_key].update(value)  # update dict
+            else:
+                d[last_key].append(value)  # value is a single value
+
+            # If list; remove duplicates and keep the last occurrence
+            if isinstance(d[last_key], Sequence):
+                d[last_key] = list(reversed(OrderedDict.fromkeys(reversed(d[last_key]))))  # reverse list to keep last occurrence instead of first