+++ tests/test_logging.py
@@ -3,6 +3,12 @@ import scanpy as sc
 import pytest
 from sctoolbox.utils import decorator as deco
 
+import sctoolbox.qc_filter as qc
+import sctoolbox.utils as utils
+
+from sctoolbox._settings import settings
+logger = settings.logger
+
 
 @pytest.fixture
 def adata():
@@ -29,12 +35,65 @@ def test_log_anndata(adata):
 
     # check if log is in adata.uns
     assert "sctoolbox" in adata.uns
-    assert "test_func" in adata.uns["sctoolbox"]
+    assert "log" in adata.uns["sctoolbox"]
+    assert "test_func" in adata.uns["sctoolbox"]["log"]
     # check if run_1 and run_2 are in adata.uns
-    assert adata.uns["sctoolbox"]["test_func"]["run_1"]['kwargs']["param1"] == 1
-    assert adata.uns["sctoolbox"]["test_func"]["run_1"]['kwargs']["param2"] is None
-    assert adata.uns["sctoolbox"]["test_func"]["run_1"]['kwargs']["param3"] == "test"
-    assert adata.uns["sctoolbox"]["test_func"]["run_1"]['kwargs']["param4"] == 1.0
-    assert adata.uns["sctoolbox"]["test_func"]["run_1"]['kwargs']["param5"] is True
+    assert adata.uns["sctoolbox"]["log"]["test_func"]["run_1"]['kwargs']["param1"] == 1
+    assert adata.uns["sctoolbox"]["log"]["test_func"]["run_1"]['kwargs']["param2"] is None
+    assert adata.uns["sctoolbox"]["log"]["test_func"]["run_1"]['kwargs']["param3"] == "test"
+    assert adata.uns["sctoolbox"]["log"]["test_func"]["run_1"]['kwargs']["param4"] == 1.0
+    assert adata.uns["sctoolbox"]["log"]["test_func"]["run_1"]['kwargs']["param5"] is True
 
-    assert adata.uns["sctoolbox"]["test_func"]["run_2"]['kwargs']["param1"] == 1
+    assert adata.uns["sctoolbox"]["log"]["test_func"]["run_2"]['kwargs']["param1"] == 1
+
+
+def test_get_parameter_table(adata):
+    """ Test if get_parameter_table works. """
+
+    # Run a few functions on the adata
+    qc.calculate_qc_metrics(adata)
+    qc.predict_sex(adata, "sample", threshold=0.1)  # threshold is kwargs
+
+    table = utils.get_parameter_table(adata)
+
+    assert table.shape[0] == 2  # two functions were run
+    assert table.loc[1, "kwargs"] == {"threshold": 0.1}  # check if kwargs are correctly stored for predict_sex
+    assert set(["func", "args", "kwargs", "user", "timestamp"]).issubset(table.columns)
+
+
+def test_add_uns_info(adata):
+    """ Test if add_uns_info works on both string and list keys. """
+
+    utils.add_uns_info(adata, "akey", "info")
+
+    assert "akey" in adata.uns["sctoolbox"]
+    assert adata.uns["sctoolbox"]["akey"] == "info"
+
+    utils.add_uns_info(adata, ["upper", "lower"], "info")
+    assert "upper" in adata.uns["sctoolbox"]
+    assert adata.uns["sctoolbox"]["upper"]["lower"] == "info"
+
+    utils.add_uns_info(adata, ["upper", "lower"], "info2", how="append")
+    assert adata.uns["sctoolbox"]["upper"]["lower"] == ["info", "info2"]
+
+
+def test_user_logging():
+    """ Test is logfile is correctly overwritten """
+
+    settings.log_file = "test.log"
+    logger.info("test_info")
+    assert "test_info" in open(settings.log_file).read()
+
+    # Set again to the same file
+    settings.log_file = "test.log"
+    logger.info("test_info2")
+    content = open(settings.log_file).read()
+    assert "test_info" in content  # check that the first log is still there
+    assert "test_info2" in content
+
+    # Set to overwrite the file
+    settings.overwrite_log = True
+    logger.info("test_info3")
+    content = open(settings.log_file).read()
+    assert "test_info2" not in content  # previous log was overwritten
+    assert "test_info3" in content      # new log is there